+++ src/telliot_core/datasource.py
@@ -11,9 +11,9 @@ from typing import List
 from typing import TypeVar
 
 from telliot_core.model.base import Base
-from telliot_core.types.datapoint import DataPoint
-from telliot_core.types.datapoint import datetime_now_utc
-from telliot_core.types.datapoint import OptionalDataPoint
+from telliot_core.dtypes.datapoint import DataPoint
+from telliot_core.dtypes.datapoint import datetime_now_utc
+from telliot_core.dtypes.datapoint import OptionalDataPoint
 
 T = TypeVar("T")
