+++ src/telliot_core/dtypes/datapoint.py
