+++ src/telliot_core/dtypes/value_type.py
