+++ src/telliot_core/queries/legacy_query.py
@@ -4,8 +4,8 @@
 from dataclasses import dataclass
 
 from telliot_core.queries.query import OracleQuery
-from telliot_core.types.float_type import UnsignedFloatType
-from telliot_core.types.value_type import ValueType
+from telliot_core.dtypes.float_type import UnsignedFloatType
+from telliot_core.dtypes.value_type import ValueType
 
 
 @dataclass
