+++ src/telliot_core/queries/string_query.py
@@ -6,7 +6,7 @@
 from dataclasses import dataclass
 
 from telliot_core.queries.query import OracleQuery
-from telliot_core.types.value_type import ValueType
+from telliot_core.dtypes.value_type import ValueType
 
 
 @dataclass
